[not verified] move hooks at block-editor/ level
