+++ projects/packages/videopress/src/client/block-editor/hooks/use-block-attributes/index.js
